refactor: config -> controller 패키지 수정
